First way on work with pickle (bad way)
