+++ users/EANabatov/work_with_files/First variant/Second_point.py
@@ -1,0 +1,4 @@
+from First_point import main_obj
+
+main_obj.add("Goshan", "84567435674", "Vladivostok")
+print(main_obj.load())
