+++ users/EANabatov/work_with_files/First variant/Third_point.py
@@ -1,0 +1,3 @@
+from Second_point import main_obj
+
+print(main_obj.load())
